[#1554] Refactor v2 models
Organize common sections together: field definitions and class-level
methods/attributes.
