+++ apps/api/v2/models/management/metadata_management.rb
@@ -1,0 +1,11 @@
+# apps/api/v2/models/management/metadata_management.rb
+
+module V2
+  class Metadata < Familia::Horreum
+
+    module Management
+    end
+
+    extend Management
+  end
+end
