+++ apps/api/v2/models/metadata.rb
@@ -3,78 +3,6 @@
 module V2
   class Metadata < Familia::Horreum
 
-    feature :safe_dump
-    feature :expiration
-
-    ttl 14.days
-    prefix :metadata
-
-    identifier :generate_id
-
-    field :custid
-    field :state
-    field :secret_key
-    field :secret_shortkey
-    field :secret_ttl
-    field :lifespan
-    field :share_domain
-    field :passphrase
-    field :viewed
-    field :received
-    field :shared
-    field :burned
-    field :created
-    field :updated
-    # NOTE: There is no `expired` timestamp field since we can calculate
-    # that based on the `secret_ttl` and the `created` timestamp. See
-    # the secret_expired? and expiration methods.
-    field :recipients
-    field :truncate # boolean
-
-    # NOTE: this field is a nullop. It's only populated if a value was entered
-    # into a hidden field which is something a regular person would not do.
-    field :token
-
-    # NOTE: Safe dump fields are loaded once at start time so they're
-    # immune to hot reloads.
-    @safe_dump_fields = [
-      { :identifier => ->(obj) { obj.identifier } },
-      :key,
-      :custid,
-      :state,
-      :secret_shortkey,
-      :secret_ttl,
-      { :metadata_ttl => ->(m) { m.lifespan } },
-      :lifespan,
-      :share_domain,
-      :created,
-      :updated,
-      :shared,
-      :received,
-      :burned,
-      :viewed,
-      :recipients,
-
-      { :shortkey => ->(m) { m.key.slice(0, 8) } },
-      { :show_recipients => ->(m) { !m.recipients.to_s.empty? } },
-
-      { :is_viewed => ->(m) { m.state?(:viewed) } },
-      { :is_received => ->(m) { m.state?(:received) } },
-      { :is_burned => ->(m) { m.state?(:burned) } },
-      { :is_expired => ->(m) { m.state?(:expired) } },
-      { :is_orphaned => ->(m) { m.state?(:orphaned) } },
-      { :is_destroyed => ->(m) { m.state?(:received) || m.state?(:burned) || m.state?(:expired) || m.state?(:orphaned) } },
-
-      # We use the hash syntax here since `:truncated?` is not a valid symbol.
-      { :is_truncated => ->(m) { m.truncated? } },
-
-      { :has_passphrase => ->(m) { m.has_passphrase? } },
-    ]
-
-    def init
-      self.state ||= 'new'
-    end
-
     def generate_id
       @key ||= Familia.generate_id.slice(0, 31)
       @key
@@ -104,8 +32,6 @@ module V2
     end
     alias :natural_ttl :natural_duration
 
-    alias :secret_expiration_in_seconds :secret_ttl
-
     def secret_expiration
       # Unix timestamp of when the secret will expire. Based on
       # the secret_ttl and the created time of the metadata
@@ -117,7 +43,6 @@ module V2
       # Colloquial representation of the TTL. e.g. "1 day"
       OT::TimeUtils.natural_duration secret_ttl.to_i if secret_ttl
     end
-    alias :secret_natural_ttl :secret_natural_duration
 
     def secret_expired?
       Time.now.utc.to_i >= (secret_expiration || 0)
@@ -254,3 +179,6 @@ module V2
     end
   end
 end
+
+require_relative 'definitions/metadata_definition'
+require_relative 'management/metadata_management'
